drivers: misc: Add support for nCipher HSM devices
From: David Kim <david.kim@ncipher.com>

This is the driver for nCipher’s Solo and Solo XC hardware security modules.
These modules implement a proprietary command set (the ’nCore API’) to
perform cryptographic operations - key generation, signature, and so on. HSM
commands and their replies are passed in a serialised binary format over the
PCIe bus via a shared memory region. Multiple commands may be in-flight at
any one time - command processing is multi-threaded and asynchronous. A write
operation may, therefore, deliver multiple commands, and multiple replies may
be retrieved in one read operation.

The HSM is supported by software supplied by nCipher. The driver supports
only a single user-space connection, so the support software includes a
daemon called ‘hardserver’ . This accepts network or Unix- domain socket
connections from client applications, and multiplexes / demultiplexes
commands and replies from these clients to one or more HSM devices.

The nCipher support software includes a variety of command-line and graphical
tools for HSM management, and a number of libraries which interface between
numerous industry-standard APIs and the nCore API commands implemented by the
HSM. On Linux, the JCE (Java Cryptography Extensions) and C-language PKCS#11
interfaces are supported.

Previously, the support software included the kernel driver as source code,
which is built against the host machine’s kernel headers when the support
software is installed.  For most standard OS distributions this isn’t a
problem, but many users in security- sensitive environments (such as banking
services) remove all unnecessary components for their own in-house
distributions. For these users, the need to support compiling and loading of
externally- supplied kernel modules is problematic, and adds extra friction
when the kernel needs to be updated.

In general, users who have an nCipher PCI HSM will be able to download the
required support software from nCipher. The devices themselves are subject to
export control restrictions under US and EU law, and because of this we are
unable to make the software freely available, or offer detailed technical
support, in all territories.

Co-developed-by: Tim Magee <tim.magee@ncipher.com>
Signed-off-by: Tim Magee <tim.magee@ncipher.com>
Signed-off-by: Dave Kim <david.kim@ncipher.com>